refactor(viewer): delete AnnotationDrawBoard component
