+++ libs/insight-viewer/src/Annotation/hooks/useAnnotationEvent/index.ts
